requirements.txt
beautifulsoup4>=4.8
mwparserfromhell>=0.5
python-box>=3.4
pywikibot>=3.0
requests>=2.22
